added registration date in users list removed language checks
formatLogin fixes
- various personal page fixes
- Added group self-registration option (Enterprise edition)
content safeguarding made optional
- Added option for compressing test results in database

removed traffic tab from branch reports
minor updates
minor updates
removed traffic tab from branch reports

git-svn-id: https://efrontlearning.svn.sourceforge.net/svnroot/efrontlearning@510 b0e902ce-d33f-0410-b758-e8627170a121

+++ community/www/themes/default/templates/includes/module_tests.tpl
@@ -1231,7 +1231,7 @@ var quickformSkillQuestCount = '{$T_QUICKTEST_FORM.skill_questions_count_row.htm
         <div class = "headerTools">
          <span>
           <img src = "images/16x16/error_delete.png" alt = "{$smarty.const._RESETEXECUTIONSFORALLUSERS}" title = "{$smarty.const._RESETEXECUTIONSFORALLUSERS}"/>
-          <a href = "javascript:void(0)" onclick = "deleteAllTestsForAllUsers(this);">{$smarty.const._RESETEXECUTIONSFORALLUSERS}</a>
+          <a href = "javascript:void(0)" onclick = "if (confirm('{$smarty.const._IRREVERSIBLEACTIONAREYOUSURE}')) deleteAllTestsForAllUsers(this);">{$smarty.const._RESETEXECUTIONSFORALLUSERS}</a>
          </span>
         </div>
                                     <table class = "sortedTable" style = "width:100%">
